HomeWork11

+++ HomeWork3/HomeWork3/Program.cs
@@ -13,14 +13,14 @@ namespace HomeWork3
             //sum =   1/1 + 1/2 + 1/3 + 1/4 + 1/5 + 1/6 + 1/7 + ..... 1/n
             int n;
 
-            double sum = 0;
+            float sum = 0;
 
             Console.WriteLine("Nhap vao gia tri n");
             n = Convert.ToInt32(Console.ReadLine());
 
             for(int i = 1; i <= n; i++)
             {
-                sum = sum + (1.0 / i); // 1.0 is float => 1.0/i is Float 
+                sum = sum + (1.0f / i); // 1.0 is float => 1.0/i is Float 
 
             }
